Restructure the code to enable resue and improve readability
Restructure the code, in order to allow the creation of a package for
bunny. Furthermore, it will help to have a better structure of the code
and improve the understandinf and readability.

However, much more work is required.
