+++ cmd/main.go
@@ -15,32 +15,25 @@
 package main
 
 import (
-	"encoding/base64"
 	"encoding/json"
 	"context"
 	"flag"
 	"os"
 	"fmt"
-	"bytes"
-	"strings"
 	"io/ioutil"
 
+	"bunny/hops"
 	"github.com/moby/buildkit/exporter/containerimage/exptypes"
 	"github.com/moby/buildkit/frontend/gateway/grpcclient"
 	"github.com/moby/buildkit/util/appcontext"
 	"github.com/moby/buildkit/client/llb"
 	"github.com/moby/buildkit/frontend/gateway/client"
 	ocispecs "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/moby/buildkit/frontend/dockerfile/instructions"
-	"github.com/moby/buildkit/frontend/dockerfile/parser"
 )
 
 const (
-	unikraftKernelPath string = "/unikraft/bin/kernel"
-	unikraftHub        string = "unikraft.org"
-	packContextName    string = "context"
+	buildContextName   string = "context"
 	clientOptFilename  string = "filename"
-	uruncJSONPath      string = "/urunc.json"
 )
 
 type CLIOpts struct {
@@ -51,12 +44,6 @@ type CLIOpts struct {
 	// Choose the execution mode. If set, then bunny will not act as a
 	// buidlkit frontend. Instead it will just print the LLB.
 	PrintLLB       bool
-}
-
-type PackInstructions struct {
-	Base   string			  // The Base image to use
-	Copies []instructions.CopyCommand // Copy commands
-	Annots map[string]string	  // Annotations
 }
 
 var version string
@@ -86,113 +73,9 @@ func parseCLIOpts() CLIOpts {
 	return opts
 }
 
-func parseFile(fileBytes []byte) (*PackInstructions, error) {
-	var instr *PackInstructions
-	instr = new(PackInstructions)
-	instr.Annots = make(map[string]string)
-
-	r := bytes.NewReader(fileBytes)
-
-	// Parse the Dockerfile
-	parseRes, err := parser.Parse(r)
-	if err != nil {
-		fmt.Printf("Failed to parse file: %v\n", err)
-		return nil, err
-	}
-
-	// Traverse Dockerfile commands
-	for _, child := range parseRes.AST.Children {
-		cmd, err := instructions.ParseInstruction(child)
-		if err != nil {
-			fmt.Printf("Failed to parse instruction %s: %v\n", child.Value, err)
-			return nil, err
-		}
-		switch c := cmd.(type) {
-		case *instructions.Stage:
-			// Handle FROM
-			if instr.Base != "" {
-				return nil, fmt.Errorf("Multi-stage builds are not supported")
-			}
-			instr.Base = c.BaseName
-		case *instructions.CopyCommand:
-			// Handle COPY
-			instr.Copies = append(instr.Copies, *c)
-		case *instructions.LabelCommand:
-			// Handle LABLE annotations
-			for _, kvp := range c.Labels {
-				annotKey := strings.Trim(kvp.Key, "\"")
-				instr.Annots[annotKey] = strings.Trim(kvp.Value, "\"")
-			}
-		case instructions.Command:
-			// Catch all other commands
-			fmt.Printf("UNsupported command%s\n", c.Name())
-		default:
-			fmt.Printf("%f is not a command type\n", c)
-		}
-
-	}
-
-	return instr, nil
-}
-
-func copyIn(base llb.State, from string, src string, dst string) llb.State {
-	var copyState llb.State
-	var localSrc llb.State
-
-	localSrc = llb.Local(packContextName)
-	copyState = base.File(llb.Copy(localSrc, src, dst, &llb.CopyInfo{
-				CreateDestPath: true,}))
-
-	return copyState
-}
-
-func constructLLB(instr PackInstructions) (*llb.Definition, error) {
-	var base llb.State
-	uruncJSON := make(map[string]string)
-
-	// Create urunc.json file, since annotations do not reach urunc
-	for annot, val := range instr.Annots {
-		encoded := base64.StdEncoding.EncodeToString([]byte(val))
-		uruncJSON[annot] = string(encoded)
-	}
-	uruncJSONBytes, err := json.Marshal(uruncJSON)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to marshal urunc json: %v", err)
-	}
-
-	// Set the base image where we will pack the unikernel
-	if instr.Base == "scratch" {
-		base = llb.Scratch()
-	} else if strings.HasPrefix(instr.Base, unikraftHub) {
-		// Define the platform to qemu/amd64 so we cna pull unikraft images
-		platform := ocispecs.Platform{
-			OS:           "qemu",
-			Architecture: "amd64",
-		}
-		base = llb.Image(instr.Base, llb.Platform(platform),)
-	} else {
-		base = llb.Image(instr.Base)
-	}
-
-	// Perform any copies inside the image
-	for _, aCopy := range instr.Copies {
-		base = copyIn(base, packContextName, aCopy.SourcePaths[0], aCopy.DestPath)
-	}
-
-	// Create the urunc.json file in the rootfs
-	base = base.File(llb.Mkfile(uruncJSONPath, 0644, uruncJSONBytes))
-
-	dt, err := base.Marshal(context.TODO(), llb.LinuxAmd64)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to marshal LLB state: %v", err)
-	}
-
-	return dt, nil
-}
-
 func readFileFromLLB(ctx context.Context, c client.Client, filename string) ([]byte, error) {
 	// Get the file from client's context
-	fileSrc := llb.Local(packContextName, llb.IncludePatterns([]string {filename}),
+	fileSrc := llb.Local(buildContextName, llb.IncludePatterns([]string {filename}),
 				llb.WithCustomName("Internal:Read-" + filename))
 	fileDef, err := fileSrc.Marshal(ctx)
 	if err != nil {
@@ -241,11 +124,11 @@ func annotateRes(annots map[string]string, res *client.Result) (*client.Result, 
 		},
 	}
 
-	uruncJSONBytes, err := json.Marshal(config)
+	imageConfig, err := json.Marshal(config)
 	if err != nil {
 		return nil, fmt.Errorf("Failed to marshal urunc json: %v", err)
 	}
-	res.AddMeta(exptypes.ExporterImageConfigKey, uruncJSONBytes)
+	res.AddMeta(exptypes.ExporterImageConfigKey, imageConfig)
 	for annot, val := range annots {
 		res.AddMeta(exptypes.AnnotationManifestKey(nil, annot), []byte(val))
 	}
@@ -256,28 +139,28 @@ func annotateRes(annots map[string]string, res *client.Result) (*client.Result, 
 
 func bunnyBuilder(ctx context.Context, c client.Client) (*client.Result, error) {
 	// Get the Build options from buildkit
-	packOpts := c.BuildOpts().Opts
+	buildOpts := c.BuildOpts().Opts
 
 	// Get the file that contains the instructions
-	packFile := packOpts[clientOptFilename]
-	if packFile == "" {
+	bunnyFile := buildOpts[clientOptFilename]
+	if bunnyFile == "" {
 		return nil, fmt.Errorf("%s: was not provided", clientOptFilename)
 	}
 
 	// Fetch and read contents of user-specified file in build context
-	fileBytes, err := readFileFromLLB(ctx, c, packFile)
+	fileBytes, err := readFileFromLLB(ctx, c, bunnyFile)
 	if err != nil {
 		return nil, fmt.Errorf("Failed to fetch and read %s: %w", clientOptFilename, err)
 	}
 
-	// Parse packing instructions
-	packInst, err := parseFile(fileBytes)
+	// Parse packaging/building instructions
+	packInst, err := hops.ParseDockerFile(fileBytes)
 	if err != nil {
-		return nil, fmt.Errorf("Error parsing packing instructions", err)
+		return nil, fmt.Errorf("Error parsing unikernel image building instructions", err)
 	}
 
-	// Create the LLB definiton
-	dt, err := constructLLB(*packInst)
+	// Create the LLB definiton of packing the final image
+	dt, err := hops.PackLLB(*packInst, buildContextName)
 	if err != nil {
 		return nil, fmt.Errorf("Failed to create LLB definition : %v\n", err)
 	}
@@ -301,7 +184,7 @@ func bunnyBuilder(ctx context.Context, c client.Client) (*client.Result, error) 
 
 func main() {
 	var cliOpts CLIOpts
-	var packInst *PackInstructions
+	var packInst *hops.PackInstructions
 
 	cliOpts = parseCLIOpts()
 
@@ -317,6 +200,7 @@ func main() {
 			fmt.Printf("Could not start grpcclient: %v\n", err)
 			os.Exit(1)
 		}
+
 		return
 	}
 
@@ -333,15 +217,15 @@ func main() {
 		os.Exit(1)
 	}
 
-	// Parse file with packaging instructions
-	packInst, err = parseFile(CntrFileContent)
+	// Parse file with packaging/building instructions
+	packInst, err = hops.ParseDockerFile(CntrFileContent)
 	if err != nil {
-		fmt.Println("Error parsing packing instructions", err)
+		fmt.Println("Error parsing unikernel image building instructions", err)
 		os.Exit(1)
 	}
 
-	// Create the LLB definition
-	dt, err := constructLLB(*packInst)
+	// Create the LLB definition of packing the final image
+	dt, err := hops.PackLLB(*packInst, buildContextName)
 	if err != nil {
 		fmt.Printf("Failed to create LLB definition : %v\n", err)
 		os.Exit(1)
